refactor: move search form logic into dedicated hooks (calendar, dates, guests, query, search)
