+++ src/shared/ui/header/search-form/hooks/useGuests.ts
@@ -1,5 +1,5 @@
-import { useState } from "react";
-import { useTranslation } from "react-i18next";
+import { useState } from 'react';
+import { useTranslation } from 'react-i18next';
 
 export const useGuests = () => {
   const { t } = useTranslation();
