+++ src/shared/ui/header/search-form/hooks/useQueryControl.ts
@@ -1,0 +1,9 @@
+import { useState } from 'react';
+import useFetchLocations from './useFetchLocations';
+
+export const useQueryControl = () => {
+  const [query, setQuery] = useState('');
+  const { suggestions, setSuggestions } = useFetchLocations(query);
+
+  return { query, setQuery, suggestions, setSuggestions };
+};
